requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filetype==1.2.0
Flask==3.1.0
Flask-DebugToolbar==0.16.0
Flask-Mailman==1.1.1
Flask-WTF==1.2.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.2
google-cloud-secret-manager==2.23.3
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.2
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
